#32 - Refactoring scenograph bootup
